pyproject.toml
[project]
name = "iprjplus-backend"
version = "0.1.0"
description = "Backend do IPRJ Plus"
requires-python = ">=3.12"

dependencies = [
  "django==5.2.1",
  "djangorestframework==3.15.1",
  "djangorestframework-simplejwt==5.3.1"
]

[dependency-groups]
dev = [
  "black>=24.4.2",
  "isort>=5.13.2",
  "ruff>=0.4.4"
]

[tool.ruff]
line-length = 129
target-version = "py312"
exclude = [
  ".venv",
  "migrations"
]
fix = true
unsafe-fixes = true
